Also update objid in replay mode.

+++ cm-replay.c
@@ -619,6 +619,6 @@ void cm_print_replay_info(void)
                  cm_coreid,
                  (uint64_t)cpu_single_env->ENVPC,
                  cm_tb_exec_cnt[cm_coreid],
-                 memop);
+                 (int)memop);
 }
